src/requirements.txt
# Face Mask Detection – Deep Learning Project
# Install with:  pip install -r requirements.txt

tensorflow>=2.12.0
tensorflow-hub>=0.14.0       # optional – required for ViT-B16 in Experiment 3
numpy>=1.23.0
matplotlib>=3.6.0
seaborn>=0.12.0
scikit-learn>=1.2.0
Pillow>=9.0.0
kaggle>=1.5.16               # optional – required for automatic dataset download
